now we keep orders with qt = zero to remember that its cancelled orders

+++ src/controller/Transaction.hx
@@ -75,6 +75,8 @@ class Transaction extends controller.Controller
 			throw Error("/member/payments/" + op.user.id, t._("Action forbidden"));		
 		}
 		
+		if (op.getPaymentType() == "lemonway-ec") throw Error("/member/payments/"+op.user.id, t._("Editing a credit card payment is not allowed"));
+		
 		op.lock();
 		
 		var f = new sugoi.form.Form("payement");
@@ -113,6 +115,7 @@ class Transaction extends controller.Controller
 	
 	public function doDelete(op:db.Operation){	
 		if (!app.user.canAccessMembership() || op.group.id != app.user.amap.id ) throw Error("/member/payments/" + op.user.id, t._("Action forbidden"));		
+		if (op.getPaymentType() == "lemonway-ec") throw Error("/member/payments/" + op.user.id, t._("Deleting a credit card payment is not allowed"));
 		
 		if (checkToken()){
 			op.delete();
